Move api tests into separate module
